requirements.txt
flask==2.3.3
openai>=1.30.0
python-dotenv==1.0.0
pydub==0.25.1
numpy>=1.24.0
sounddevice==0.4.6
soundfile==0.12.1
requests>=2.31.0
python-multipart==0.0.6
gunicorn==21.2.0
